requirements.txt
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.122.0
google-generativeai==0.5.0
python-dotenv==1.0.1
google-cloud-secret-manager==2.20.0
